+++ src/queue-interruption.js
@@ -7,6 +7,8 @@ module.exports.pause = function interruptionPause (q) {
   logger(`pause`)
   return new Promise((resolve, reject) => {
     q._paused = true
+    logger(`Event: pausing [${q.id}]`)
+    q.emit(enums.status.pausing, q.id)
     if (q.running < 1) { return resolve() }
     let intId = setInterval(function pausing () {
       logger(`Pausing, waiting on running jobs: [${q.running}]`)